database connection

+++ PaymentGatewayPublishedLanguage/Events/AccountUpdated.cs
@@ -7,9 +7,9 @@ namespace PaymentGateway.PublishedLanguage.Events
     {
         public string IbanCode { get; set; }
         public DateTime Date { get; set; }
-        public double Amount { get; set; }
+        public decimal Amount { get; set; }
 
-        public AccountUpdated(string ibanCode, DateTime date, double amount)
+        public AccountUpdated(string ibanCode, DateTime date, decimal amount)
         {
             IbanCode = ibanCode;
             Date = date;
