+++ PaymentGatewayPublishedLanguage/Events/TransactionCreated.cs
@@ -5,11 +5,11 @@ namespace PaymentGateway.PublishedLanguage.Events
 {
     public class TransactionCreated : INotification
     {
-        public double Amount { get; set; }
+        public decimal Amount { get; set; }
         public string Currency { get; set; }
         public DateTime Date { get; set; }
 
-        public TransactionCreated(double amount, string currency, DateTime date)
+        public TransactionCreated(decimal amount, string currency, DateTime date)
         {
             Amount = amount;
             Currency = currency;
